# searcher: optimize capture groups