--- a/cmd/searcher/internal/search/search_regex_test.go
+++ b/cmd/searcher/internal/search/search_regex_test.go
@@ -93,6 +93,18 @@ func BenchmarkSearchRegex_large_re_anchor(b *testing.B) {
 	})
 }
 
+func BenchmarkSearchRegex_large_capture_group(b *testing.B) {
+	benchSearchRegex(b, &protocol.Request{
+		Repo:   "github.com/golang/go",
+		Commit: "0ebaca6ba27534add5930a95acffa9acff182e2b",
+		PatternInfo: protocol.PatternInfo{
+			Pattern:         "(TODO|FIXME)",
+			IsRegExp:        true,
+			IsCaseSensitive: true,
+		},
+	})
+}
+
 func BenchmarkSearchRegex_large_path(b *testing.B) {
 	do := func(b *testing.B, content, path bool) {
 		benchSearchRegex(b, &protocol.Request{
@@ -162,6 +174,18 @@ func BenchmarkSearchRegex_small_re_anchor(b *testing.B) {
 		Commit: "4193810334683f87b8ed5d896aa4753f0dfcdf20",
 		PatternInfo: protocol.PatternInfo{
 			Pattern:         "^func +[A-Z]",
+			IsRegExp:        true,
+			IsCaseSensitive: true,
+		},
+	})
+}
+
+func BenchmarkSearchRegex_small_capture_group(b *testing.B) {
+	benchSearchRegex(b, &protocol.Request{
+		Repo:   "github.com/sourcegraph/go-langserver",
+		Commit: "4193810334683f87b8ed5d896aa4753f0dfcdf20",
+		PatternInfo: protocol.PatternInfo{
+			Pattern:         "(TODO|FIXME)",
 			IsRegExp:        true,
 			IsCaseSensitive: true,
 		},